split utisl
