+++ src/assertions.ts
@@ -44,7 +44,8 @@ import {
   AssertionValue,
   ProviderResponse,
 } from './types';
-import { transformOutput, getNunjucksEngine, extractJsonObjects } from './util';
+import { transformOutput, extractJsonObjects } from './util';
+import { getNunjucksEngine } from './util/templates';
 
 const ASSERTIONS_MAX_CONCURRENCY = process.env.PROMPTFOO_ASSERTIONS_MAX_CONCURRENCY
   ? parseInt(process.env.PROMPTFOO_ASSERTIONS_MAX_CONCURRENCY, 10)
