+++ src/onboarding.ts
@@ -3,7 +3,7 @@ import fs from 'fs';
 import inquirer from 'inquirer';
 import path from 'path';
 import logger from './logger';
-import { getNunjucksEngine } from './util';
+import { getNunjucksEngine } from './util/templates';
 
 export const CONFIG_TEMPLATE = `# Learn more about building a configuration: https://promptfoo.dev/docs/configuration/guide
 description: "My eval"
