+++ src/redteam/index.ts
@@ -4,8 +4,8 @@ import invariant from 'tiny-invariant';
 import logger from '../logger';
 import { loadApiProvider } from '../providers';
 import type { ApiProvider, TestCase, TestSuite } from '../types';
+import { extractVariablesFromTemplates } from '../util/templates';
 import { REDTEAM_MODEL, ALL_PLUGINS, DEFAULT_PLUGINS } from './constants';
-import { extractVariablesFromTemplates } from './injectVar';
 import { addInjections } from './methods/injections';
 import { addIterativeJailbreaks } from './methods/iterative';
 import CompetitorPlugin from './plugins/competitors';
