--- src/redteam/injectVar.ts
@@ -1,46 +1,0 @@
-/**
- * Parse Nunjucks template to extract variables.
- * @param template - The Nunjucks template string.
- * @returns An array of variables used in the template.
- */
-export function extractVariablesFromTemplate(template: string): string[] {
-  const variableSet = new Set<string>();
-  const regex =
-    /\{\{[\s]*([^{}\s|]+)[\s]*(?:\|[^}]+)?\}\}|\{%[\s]*(?:if|for)[\s]+([^{}\s]+)[\s]*.*?%\}/g;
-  const commentRegex = /\{#[\s\S]*?#\}/g;
-
-  // Remove comments
-  template = template.replace(commentRegex, '');
-
-  let match;
-  while ((match = regex.exec(template)) !== null) {
-    const variable = match[1] || match[2];
-    if (variable) {
-      // Split by dot and add only the full path
-      variableSet.add(variable);
-    }
-  }
-
-  // Handle for loops separately
-  const forLoopRegex = /\{%[\s]*for[\s]+(\w+)[\s]+in[\s]+(\w+)[\s]*%\}/g;
-  while ((match = forLoopRegex.exec(template)) !== null) {
-    variableSet.delete(match[1]); // Remove loop variable
-    variableSet.add(match[2]); // Add the iterated variable
-  }
-
-  return Array.from(variableSet);
-}
-
-/**
- * Extract variables from multiple Nunjucks templates.
- * @param templates - An array of Nunjucks template strings.
- * @returns An array of variables used in the templates.
- */
-export function extractVariablesFromTemplates(templates: string[]): string[] {
-  const variableSet = new Set<string>();
-  for (const template of templates) {
-    const variables = extractVariablesFromTemplate(template);
-    variables.forEach((variable) => variableSet.add(variable));
-  }
-  return Array.from(variableSet);
-}
