+++ src/redteam/providers/iterative.ts
@@ -1,7 +1,7 @@
 import dedent from 'dedent';
 import invariant from 'tiny-invariant';
 import type { ApiProvider, CallApiContextParams, CallApiOptionsParams } from '../../types';
-import { getNunjucksEngine } from '../../util';
+import { getNunjucksEngine } from '../../util/templates';
 
 const ATTACKER_MODEL = 'gpt-4o';
