+++ src/util/index.ts
@@ -49,27 +49,9 @@ import {
   isProviderOptions,
   OutputFileExtension,
 } from '../types';
+import { getNunjucksEngine } from './templates';
 
 const DEFAULT_QUERY_LIMIT = 100;
-
-export function getNunjucksEngine(filters?: NunjucksFilterMap) {
-  if (process.env.PROMPTFOO_DISABLE_TEMPLATING) {
-    return {
-      renderString: (template: string) => template,
-    };
-  }
-
-  const env = nunjucks.configure({
-    autoescape: false,
-  });
-
-  if (filters) {
-    for (const [name, filter] of Object.entries(filters)) {
-      env.addFilter(name, filter);
-    }
-  }
-  return env;
-}
 
 export async function writeOutput(
   outputPath: string,
